+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## Master
+
+### Enhancements
+
+- The output format `application/vnd.refract.parse-result+json; version=0.6`
+  and `application/vnd.refract.parse-result+yaml; version=0.6` were added which
+  allows you to serialise to the Refract 0.6 JSON/YAML Serialisation format.
+
 ## 0.3.0 - 2017-06-11
 
 - Update to Fury 3.0.0-beta.3 which updates Refract JSON serialisation rules to